baymax_project-main/backend/requirements.txt
flask
flask-cors
psycopg2-binary
# Add any other dependencies below
